move (unused) idb-shim to test/util
