+++ test/test.js
@@ -3,7 +3,7 @@ var leveljs = require('../')
 var testCommon = require('./testCommon')
 
 // load IndexedDBShim in the tests
-// require('./idb-shim.js')()
+// require('./util/idb-shim.js')()
 
 /*** compatibility with basic LevelDOWN API ***/
 require('abstract-leveldown/abstract/leveldown-test').args(leveljs, tape, testCommon)
